+++ plugins/ameliabooking/src/Infrastructure/WP/InstallActions/DB/Booking/EventsTagsTable.php
@@ -28,7 +28,7 @@ class EventsTagsTable extends AbstractDatabaseTable
 
         return "CREATE TABLE {$table} (
                    `id` INT(11) NOT NULL AUTO_INCREMENT,
-                   `eventId` INT(11) NOT NULL,
+                   `eventId` bigint(20) NOT NULL,
                    `name` varchar({$name}) NOT NULL,
                     PRIMARY KEY (`id`)
                 ) DEFAULT CHARSET=utf8 COLLATE utf8_general_ci";
